requirements.txt
langchain== 0.0.343
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
streamlit-extras
altair==4.1.0
openai==1.3.6
tiktoken
flask
scikit-learn
pinecone-client
sqlalchemy
pyodbc